// teeok-loyixasi/src/lib/enums/product.enum.ts
export enum ProductCollection {
  TTEOK = "TTEOK",
  HANGWA = "HANGWA",
  SIKHYE = "SIKHYE",
  SUJEONGGWA = "SUJEONGGWA",
  OMIJA = "OMIJA",
  SET = "SET",
}

export enum ProductSize {
  SMALL = "SMALL",
  NORMAL = "NORMAL",
  LARGE = "LARGE",
  FAMILY = "FAMILY",
}

export enum ProductStatus {
  PAUSE = "PAUSE",
  PROCESS = "PROCESS",
  DELETE = "DELETE",
}